Finish chapter 06
If anything, the troubles I had getting this working should serve as a
reminder that I should always read examples and documentation properly
instead of skimming through it.

There were two problems that managed to cripple my progress on porting
this tutorial.

1: Shader reflection doesn't seem to remember that there's a uniform in
the vertex shader once I add uniforms to the fragment shader
(amethyst/rendy#220). This is remedied by specifying the uniform
bindings manually.

2: The image I was trying to render onto the triangle wouldn't upload
properly, leaving me with an entirely transparent black texture. The
root cause of this issue was that I forgot to call `factory.maintain()`
before running the graph (or at all, really).